+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version 0.6-6.9006 (2019-01-21)
+* Improvements on `plotCor` to deal with long column names in the correlation matrix. Long column names (> 5 
+  characters) are now replaced with x1, x2, ..., xn, where n is the number of columns in the correlation matrix.
+  A message is issues identifying coluns and their respective index.
+
 # Version 0.6-6.9005 (2019-01-15)
 * Adds `plotCor`, a function to create correlation plots using a colorblind-friendly pallete.